UserList gets two parameters: Users and Type
