+++ DummyData.ts
@@ -1,0 +1,25 @@
+export const dummyData = [
+    {
+      id: "1",
+      username: "top_G_3",
+      nickname: "Aleks069",
+      avatarUrl: require("./assets/images/Max.jpeg"),
+      posts: [
+        require("./assets/images/Ei.jpeg"),
+        require("./assets/images/Luca.jpeg"),
+      ],
+      isFollowing: true,
+    },
+    {
+      id: "2",
+      username: "koenig_kemal",
+      nickname: "Kevin",
+      avatarUrl: require("./assets/images/Kevin.jpeg"),
+      posts: [
+        require("./assets/images/Adrian.jpeg"),
+        require("./assets/images/Luca.jpeg"),
+        require("./assets/images/Aleks.jpeg"),
+      ],
+      isFollowing: true,
+    },
+  ];
